+++ tests/testthat/test-bmab_gi_multiple.R
@@ -7,6 +7,9 @@ test_that("bmab_gi_multiple() arg checks work", {
   expect_false(check_start_args(alpha_start = 1, Sigma_start = 1))
 
   expect_error(bmab_gi_multiple(alpha_start = 1, gamma = 0.5, N = 20, num_actions = 2), "Incorrect number")
+  expect_error(bmab_gi_multiple(1, 1, gamma = 0.5, N = 20, num_actions = 1.9),
+               "`num_actions` must be at least 2")
+
 })
 
 test_that("bmab_gi_multiple() gives correct output", {